Render cache, Type1 parser, LCD subpixel AA, and progress improvements
- Add render-cache.hh/cc: LRU-evicted byte-size-bounded cache (128MB default)
  for rendered tiles, glyphs, and images. Thread-safe with stats API.
- Add shared-font-cache.hh/cc: process-wide font cache singleton.
- Remove per-page glyph_bitmap_cache_.clear() — glyphs persist across pages
  since font data is immutable per PDF; FIFO eviction keeps memory bounded.
- Add tiling pattern render cache in apply_tiling_pattern() — skips expensive
  sub-render (canvas/scene/content parse) on cache hit keyed by content hash,
  tile dimensions, and paint type.
- Add image decoding cache in XObject Do handler — skips parse_image_xobject
  + color conversion on cache hit keyed by (obj_num, obj_gen). Added
  CacheARGB color space type and last_image_argb_ member.
- Add glyph ARGB render cache in draw_glyph_bitmap_by_index() — skips
  per-pixel color+alpha fill loop on cache hit keyed by font/glyph/size/LCD/color.
- Add LCD subpixel anti-aliasing (A1): lcd_subpixel option in RenderOptions,
  3x oversampling with per-channel alpha extraction, pre-multiplied ARGB.
- Add Type1 CharStrings parser (A4): eexec decryption (key=55665),
  CharStrings/Subrs extraction, Type1 interpreter (26 operators),
  t1_glyph_outline() public API, t1_binary_decrypt/encrypt.
- Add render progress callback granularity (D3): weighted progress for
  Image Do (pixel-area based) and Form Do (recursive operator count).
- Add font-unicode-map.hh: Adobe glyph name to Unicode mapping.
- Add PDF writer enhancements: page builder, form manipulation, annotations.
- Update CMakeLists.txt: add render-cache.cc to build.

+++ src/blend2d-backend.hh
@@ -188,7 +188,7 @@ private:
     float char_spacing{0.0f};        // Tc - character spacing
     float word_spacing{0.0f};        // Tw - word spacing
     float horiz_scaling{100.0f};     // Tz - horizontal scaling (percentage)
-    int text_rise{0};                // Ts - text rise (baseline shift)
+    float text_rise{0.0f};           // Ts - text rise (baseline shift)
     int text_render_mode{0};         // Tr - text rendering mode (0-7)
 
     // Transformation matrix [a b c d e f]
